Remove Test Header in randomization
