--- networkit/cpp/randomization/test/CurveballGlobalTradeGeneratorGTest.h
@@ -1,23 +1,0 @@
-/*
- * CurveballGlobalTradeGeneratorGTest.h
- *
- *  Created on: Jul 11, 2017
- *      Author: Hung Tran <htran@ae.cs.uni-frankfurt.de>
- */
-
-#ifndef RANDOMIZATION_TEST_CURVEBALL_GLOBAL_TRADE_GENERATOR_H
-#define RANDOMIZATION_TEST_CURVEBALL_GLOBAL_TRADE_GENERATOR_H
-
-#include <gtest/gtest.h>
-
-namespace NetworKit {
-
-class CurveballGlobalTradeGeneratorGTest : public testing::Test {
-public:
-    CurveballGlobalTradeGeneratorGTest() = default;
-    virtual ~CurveballGlobalTradeGeneratorGTest() = default;
-};
-
-}
-
-#endif // RANDOMIZATION_TEST_CURVEBALL_GLOBAL_TRADE_GENERATOR_H
