--- networkit/cpp/randomization/test/CurveballImplGTest.h
@@ -1,19 +1,0 @@
-/*
- * CurveballImplGTest.h
- *
- *  Created on: 26.08.2018
- *      Author:  Hung Tran <htran@ae.cs.uni-frankfurt.de>
- */
-
-#ifndef RANDOMIZATION_TEST_CURVEBALL_IMPL_H
-#define RANDOMIZATION_TEST_CURVEBALL_IMPL_H
-
-#include <gtest/gtest.h>
-
-class CurveballImplGTest : public ::testing::Test {
-public:
-	CurveballImplGTest() = default;
-	virtual ~CurveballImplGTest() = default;
-};
-
-#endif // RANDOMIZATION_TEST_CURVEBALL_IMPL_H
