--- networkit/cpp/randomization/test/CurveballUniformTradeGeneratorGTest.h
@@ -1,23 +1,0 @@
-/*
- * CurveballUniformTradeGeneratorGTest.h
- *
- *  Created on: 26.05.2018
- *      Author:  Hung Tran <htran@ae.cs.uni-frankfurt.de>
- */
-
-#ifndef RANDOMIZATION_TEST_CURVEBALL_UNIFORM_TRADE_GENERATOR_H
-#define RANDOMIZATION_TEST_CURVEBALL_UNIFORM_TRADE_GENERATOR_H
-
-#include <gtest/gtest.h>
-
-namespace NetworKit {
-
-class CurveballUniformTradeGeneratorGTest : public ::testing::Test {
-public:
-    CurveballUniformTradeGeneratorGTest() = default;
-    virtual ~CurveballUniformTradeGeneratorGTest() = default;
-};
-
-}
-
-#endif // RANDOMIZATION_TEST_CURVEBALL_UNIFORM_TRADE_GENERATOR_H
